Implemented behavior for first simple tests
The class is now capable to display the first four deck cards in their
given order and to display and click a hand card. The click notifies the
registered observers. It is not possible to display a specific users
deck or to show the score due to failing class implementations.