Made loading pictures from cards faster
By loading all possible card pictures (36 in total) at the start of the
program into a map, retrieving them is a lot faster. The pictures from
the map are not copied by reference but by making a deep copy of them.
For this purpose a new static method has been added to the class.